pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "sueun-dev",email = "sueun.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "requests-html (>=0.10.0,<0.11.0)",
    "lxml-html-clean (>=0.4.1,<0.5.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"
pytest = "^8.3.4"
pyright = "^1.1.393"

